cleanup: separte sf code
