+++ src/meta/meta.h
@@ -4,6 +4,7 @@
 #include "../vgmstream.h"
 #include "../util/reader_sf.h"
 #include "../util/reader_text.h"
+#include "../util/sf_utils.h"
 
 typedef VGMSTREAM* (*init_vgmstream_t)(STREAMFILE* sf);
