+++ src/streamfile.h
@@ -130,15 +130,6 @@ STREAMFILE* open_multifile_streamfile_f(STREAMFILE** sfs, size_t sfs_size);
  * Just a wrapper, to avoid having to access the STREAMFILE's callbacks directly. */
 STREAMFILE* open_streamfile(STREAMFILE* sf, const char* pathname);
 
-/* Opens a STREAMFILE from a base pathname + new extension
- * Can be used to get companion headers. */
-STREAMFILE* open_streamfile_by_ext(STREAMFILE* sf, const char* ext);
-
-/* Opens a STREAMFILE from a base path + new filename.
- * Can be used to get companion files. Relative paths like
- * './filename', '../filename', 'dir/filename' also work. */
-STREAMFILE* open_streamfile_by_filename(STREAMFILE* sf, const char* filename);
-
 /* Reopen a STREAMFILE with a different buffer size, for fine-tuned bigfile parsing.
  * Uses default buffer size when buffer_size is 0 */
 STREAMFILE* reopen_streamfile(STREAMFILE* sf, size_t buffer_size);
@@ -161,18 +152,6 @@ static inline size_t get_streamfile_size(STREAMFILE* sf) {
 }
 
 
-/* various STREAMFILE helpers functions */
-
-/* Checks if the stream filename is one of the extensions (comma-separated, ex. "adx" or "adx,aix").
- * Empty is ok to accept files without extension ("", "adx,,aix"). Returns 0 on failure */
-int check_extensions(STREAMFILE* sf, const char* cmp_exts);
-
-/* filename helpers */
-void get_streamfile_name(STREAMFILE* sf, char* buf, size_t size);
-void get_streamfile_filename(STREAMFILE* sf, char* buf, size_t size);
-void get_streamfile_basename(STREAMFILE* sf, char* buf, size_t size);
-void get_streamfile_path(STREAMFILE* sf, char* buf, size_t size);
-void get_streamfile_ext(STREAMFILE* sf, char* buf, size_t size);
-
 void dump_streamfile(STREAMFILE* sf, int num);
+
 #endif
