+++ src/util/companion_files.c
@@ -1,7 +1,8 @@
 #include "companion_files.h"
 #include "paths.h"
 #include "../vgmstream.h"
-#include "../util/reader_text.h"
+#include "reader_text.h"
+#include "sf_utils.h"
 
 
 size_t read_key_file(uint8_t* buf, size_t buf_size, STREAMFILE* sf) {
